+++ test/TestQINToken.sol
@@ -9,7 +9,7 @@ import "../contracts/libs/SafeMath.sol";
  *  @author WorldRapidFinance <info@worldrapidfinance.com>
  */
 contract TestQINToken {
-    using SafeMath for uint256;
+    using SafeMath for uint;
 
     uint decimalMultiplier = 10**18;
